+++ opr/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 """OPR modules."""
 from .opr_param import OPR_VERSION
+from .opr_obj import OPR
 
 __version__ = OPR_VERSION