+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 ### Changed
+## [0.1] - 2024-06-05
+### Added
+- Molecular Weight Calculation
+- Sequence Validation
+- OPR class
 
 [Unreleased]: https://github.com/openscilab/opr/compare/v0.1...dev
 [0.1]: https://github.com/openscilab/opr/compare/3598e8b...v0.1